requirements.txt
flask==3.0.0
tensorflow==2.16.1
pillow==10.2.0
numpy==1.26.4
werkzeug==3.0.1
google-generativeai>=0.8.3
python-dotenv==1.0.1
